+++ NEWS.md
@@ -1,1 +1,18 @@
-First version of cliqueMS.
+0.2.3
+
+Function "computeCliques" corrected, there was a
+small bug in the "step" variable.
+Function "getIsotopes" corrected, bug when
+"filterIso" leave no isotopes in the group.
+In function "filterIso", small bug corrected.
+maxGrade parameter was not working correctly, fixed
+for function "getIsotopes".
+
+0.2.2
+Function "getIsotopes" updated: there was a
+bug when the sample had no isotopes, now corrected.
+Function "getAnnotation" updated: there was a
+bug when the sample had no isotopes, now corrected.
+
+0.2.1
+First version of cliqueMS